CBD-5667: Check archive_dir exists before modifying/removing
Reviewed-on: https://review.couchbase.org/c/build/+/199656
Tested-by: Blair Watt <blair.watt@couchbase.com>

+++ scripts/jenkins/commit_validation/archive_core_files.sh
@@ -19,8 +19,10 @@ shift
 
 # Wipe out any previously existing archive directory (so any old
 # archive doesn't get detected as an artifact by Jenkins).
-chmod -R +w $archive_dir
-rm -fr $archive_dir
+if [ -e $archive_dir ]; then
+    chmod -R +w $archive_dir
+    rm -fr $archive_dir
+fi
 
 if [ "$#" -ne 0 ]; then
     cat <<EOF
